fix(video): encode fps = frame_count / max_time so mp4 duration matches sim
Previously the paint encode was hard-coded at --fps 30. With
--drake-frames-stride=3 and dt_ctrl≈0.075s, a 30-second sim produces
≈134 frames — encoding those at 30 fps yielded a 4.5 s mp4 (6.7×
sped up), which was hard to correlate with the sim log.

Now the paint invocation computes fps = frame_count / max_time so
playback duration matches sim duration (1× real-time). Verified on
the 2026-07-18 hysteresis-fix run: 134 frames / 30 s → 4.467 fps →
mp4 duration 29.998 s.

paint_full_hud.py and paint_mode_text.py accept --fps as float now
(previously int, would have rejected 4.467).

+++ tools/visualizer/paint_full_hud.py
@@ -196,7 +196,7 @@ def main():
     ap.add_argument("--frames-dir", required=True, type=Path)
     ap.add_argument("--log-path", required=True, type=Path)
     ap.add_argument("--output", required=True, type=Path)
-    ap.add_argument("--fps", type=int, default=30)
+    ap.add_argument("--fps", type=float, default=30.0)
     args = ap.parse_args()
 
     print(f"[HUD] parsing {args.log_path}")
